# api-docs/api-doc.yml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:61988
  description: Generated server url
...
paths:
  /{version}/merchants/{merchantId}/products/{productId}:
    get:
      tags:
      - Merchant Products
      summary: Get a single active product
      operationId: get
      parameters:
      - name: merchantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: productId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: branchId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductResponse"
    put:
      tags:
      - Merchant Products
      summary: "Update a product's name, price, and image"
      operationId: update
      parameters:
      - name: merchantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: productId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                branchId:
                  type: string
                unitPrice:
                  type: string
                description:
                  type: string
                image:
                  type: string
                  format: binary
              required:
              - branchId
              - name
              - unitPrice
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductResponse"
    delete:
      tags:
      - Merchant Products
      summary: Soft-delete a product (sets active = false)
      operationId: deactivate
      parameters:
      - name: merchantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: productId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: branchId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
...
  /{version}/merchants/{merchantId}/branches/{branchId}/image:
    put:
      tags:
      - Merchant Branches
      summary: Set branch image
      operationId: setImage
      parameters:
      - name: merchantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: branchId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetBranchImageRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BranchResponse"
...
  /{version}/merchants/{id}/logo:
    put:
      tags:
      - Merchant Onboarding
      summary: Set merchant logo
      description: "Records the S3 storage key of a logo already uploaded by the client.\
        \ Expected key format: merchant/{applicationId}/logo.{ext}"
      operationId: setLogo
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetLogoRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MerchantSummary"
  /{version}/merchants/{id}:
    get:
      tags:
      - Merchant Onboarding
      summary: Get merchant application details
      operationId: get_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MerchantDetailsResponse"
    put:
      tags:
      - Merchant Onboarding
      summary: Update editable merchant details and optional status
      operationId: update_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMerchantRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MerchantDetailsResponse"
...
  /{version}/merchants/{id}/activate:
    post:
      tags:
      - Merchant Onboarding
      summary: Directly activate a merchant application (skips onboarding flow)
      operationId: activate
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MerchantSummary"
...
components:
  schemas:
    ProductResponse:
      type: object
      properties:
        productId:
          type: string
          format: uuid
        branchId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        unitPrice:
          type: number
        imageUrl:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
      - active
      - branchId
      - createdAt
      - name
      - productId
      - unitPrice
      - updatedAt
...
    SetBranchImageRequest:
      type: object
      properties:
        storageKey:
          type: string
      required:
      - storageKey
    BranchResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        merchantId:
          type: string
          format: uuid
        name:
          type: string
        active:
          type: boolean
        addressLine1:
          type: string
        addressLine2:
          type: string
        city:
          type: string
        postalCode:
          type: string
        country:
          type: string
        contactFullName:
          type: string
        contactEmail:
          type: string
        contactPhoneNumber:
          type: string
        activePaymentChannels:
          type: array
          items:
            type: string
            enum:
            - PAYPAL
            - STRIPE
            - TRUELAYER
          uniqueItems: true
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        geofenceRadiusMeters:
          type: integer
          format: int32
        googlePlaceId:
          type: string
        placeFormattedAddress:
          type: string
        placeRating:
          type: number
          format: double
        placeReviewCount:
          type: integer
          format: int32
        placeEnrichmentStatus:
          type: string
        placeEnrichedAt:
          type: string
          format: date-time
        imageUrl:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
      - active
      - activePaymentChannels
      - createdAt
      - id
      - merchantId
      - name
      - updatedAt
...
    SetLogoRequest:
      type: object
      properties:
        storageKey:
          type: string
      required:
      - storageKey
    MerchantSummary:
      type: object
      properties:
        applicationId:
          type: string
          format: uuid
        applicationReference:
          type: string
        registrationNumber:
          type: string
        status:
          type: string
          enum:
          - DRAFT
          - SUBMITTED
          - VERIFICATION
          - SCREENING
          - RISK_REVIEW
          - DECISION_PENDING
          - SETUP
          - ACTIVE
          - MONITORING
          - ACTION_REQUIRED
          - REJECTED
        submittedAt:
          type: string
          format: date-time
        lastUpdatedAt:
          type: string
          format: date-time
        logoStorageKey:
          type: string
      required:
      - applicationId
      - applicationReference
      - lastUpdatedAt
      - registrationNumber
      - status
    UpdateMerchantRequest:
      type: object
      properties:
        legalBusinessName:
          type: string
        primaryBusinessAddress:
          type: string
        contactFullName:
          type: string
        contactEmail:
          type: string
        contactPhoneNumber:
          type: string
        activePaymentChannels:
          type: array
          items:
            type: string
            enum:
            - PAYPAL
            - STRIPE
            - TRUELAYER
          uniqueItems: true
        status:
          type: string
          enum:
          - DRAFT
          - SUBMITTED
          - VERIFICATION
          - SCREENING
          - RISK_REVIEW
          - DECISION_PENDING
          - SETUP
          - ACTIVE
          - MONITORING
          - ACTION_REQUIRED
          - REJECTED
      required:
      - activePaymentChannels
      - contactEmail
      - contactFullName
      - contactPhoneNumber
      - legalBusinessName
      - primaryBusinessAddress
    MerchantDetailsResponse:
      type: object
      properties:
        applicationId:
          type: string
          format: uuid
        applicationReference:
          type: string
        registrationNumber:
          type: string
        businessType:
          type: string
        operatingCountry:
          type: string
        legalBusinessName:
          type: string
        primaryBusinessAddress:
          type: string
        contactFullName:
          type: string
        contactEmail:
          type: string
        contactPhoneNumber:
          type: string
        activePaymentChannels:
          type: array
          items:
            type: string
            enum:
            - PAYPAL
            - STRIPE
            - TRUELAYER
          uniqueItems: true
        status:
          type: string
          enum:
          - DRAFT
          - SUBMITTED
          - VERIFICATION
          - SCREENING
          - RISK_REVIEW
          - DECISION_PENDING
          - SETUP
          - ACTIVE
          - MONITORING
          - ACTION_REQUIRED
          - REJECTED
        submittedAt:
          type: string
          format: date-time
        lastUpdatedAt:
          type: string
          format: date-time
        logoStorageKey:
          type: string
        logoUrl:
          type: string
      required:
      - activePaymentChannels
      - applicationId
      - applicationReference
      - businessType
      - contactEmail
      - contactFullName
      - contactPhoneNumber
      - lastUpdatedAt
      - legalBusinessName
      - operatingCountry
      - primaryBusinessAddress
      - registrationNumber
      - status
...