# src/main/kotlin/com/elegant/software/blitzpay/merchant/api/MerchantBranchModels.kt
package com.elegant.software.blitzpay.merchant.api

import com.elegant.software.blitzpay.merchant.domain.MerchantPaymentChannel
import java.time.Instant
import java.util.UUID

data class CreateBranchRequest(
    val name: String,
    val addressLine1: String? = null,
    val addressLine2: String? = null,
    val city: String? = null,
    val postalCode: String? = null,
    val country: String? = null,
    val contactFullName: String? = null,
    val contactEmail: String? = null,
    val contactPhoneNumber: String? = null,
    val activePaymentChannels: Set<MerchantPaymentChannel> = emptySet(),
    val latitude: Double? = null,
    val longitude: Double? = null,
    val geofenceRadiusMeters: Int? = null,
    val googlePlaceId: String? = null,
)

data class UpdateBranchRequest(
    val name: String,
    val active: Boolean = true,
    val addressLine1: String? = null,
    val addressLine2: String? = null,
    val city: String? = null,
    val postalCode: String? = null,
    val country: String? = null,
    val contactFullName: String? = null,
    val contactEmail: String? = null,
    val contactPhoneNumber: String? = null,
    val activePaymentChannels: Set<MerchantPaymentChannel> = emptySet(),
    val latitude: Double? = null,
    val longitude: Double? = null,
    val geofenceRadiusMeters: Int? = null,
    val googlePlaceId: String? = null,
)

data class BranchResponse(
    val id: UUID,
    val merchantId: UUID,
    val name: String,
    val active: Boolean,
    val addressLine1: String?,
    val addressLine2: String?,
    val city: String?,
    val postalCode: String?,
    val country: String?,
    val contactFullName: String?,
    val contactEmail: String?,
    val contactPhoneNumber: String?,
    val activePaymentChannels: Set<MerchantPaymentChannel>,
    val latitude: Double?,
    val longitude: Double?,
    val geofenceRadiusMeters: Int?,
    val googlePlaceId: String?,
    val placeFormattedAddress: String?,
    val placeRating: Double?,
    val placeReviewCount: Int?,
    val placeEnrichmentStatus: String?,
    val placeEnrichedAt: Instant?,
    val imageUrl: String?,
    val createdAt: Instant,
    val updatedAt: Instant,
)